Rename Parameter "featureName" To "remove-feature" To Keep Consistency
First of all, TypeScript will complain if I use an identifier like
`feature-name`. The workaround is to declare the identifier sorrounded
by quotes i.e. `"feature-name"`.

Now, Atomist won't like this identifier and for some reason it will not
capture the value of it if you use `-` to separate the words of the of
the option. To workaround this, you should use the `.` character.

I found this while looking at the source code of the following file:

https://github.com/atomist/sdm-local/blob/master/lib/cli/invocation/command/support/runCommandOnColocatedAutomationClient.ts#L220

As you can see, for these kind of multiple word arguments they use that
character e.g. `--target-owner`